fixed building errors implement from trait for anyWrapper with Fee
