+++ modules/js-helpers/src/signer.rs
@@ -1,6 +1,6 @@
 use cosmos_sdk_proto::cosmos::tx::v1beta1::{BroadcastMode, Fee};
 use crw_client::{client::get_node_info, client::ChainClient};
-use crw_types::msg::{AnyWrapper, Msg};
+use crw_types::{any_wrapper::AnyWrapper, msg::Msg};
 use crw_wallet::{crypto::Wallet, crypto::WalletJS};
 use prost_types::Any;
 use wasm_bindgen::prelude::*;
@@ -19,9 +19,10 @@ pub async fn sign_and_send_msg(
     // convert all the js values to actual rust types
     let wallet_js: WalletJS = js_wallet.into_serde().unwrap();
     let wallet = Wallet::from(wallet_js);
-    let msg_wrapped: AnyWrapper = js_msg.into_serde().unwrap();
-    let msg: Msg = Msg(Any::from(msg_wrapped));
-    let fees: Fee = js_fees.into_serde().unwrap();
+    let wrapped_msg: AnyWrapper = js_msg.into_serde().unwrap();
+    let msg: Msg = Msg(Any::from(wrapped_msg));
+    let wrapped_fee: AnyWrapper = js_fees.into_serde().unwrap();
+    let fees: Fee = Fee::from(wrapped_fee);
 
     let response = get_node_info(lcd_addr.to_string())
         .await
