+++ packages/crw-types/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod any_wrapper;
 pub mod error;
 pub mod msg;
