+++ packages/crw-types/src/msg.rs
@@ -1,12 +1,5 @@
 //! Transaction Message representation
 use prost_types::Any;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct AnyWrapper {
-    pub type_url: String,
-    pub value: Vec<u8>,
-}
 
 /// Transaction message wrapper
 pub struct Msg(pub Any);
@@ -29,14 +22,5 @@ impl From<Any> for Msg {
 impl From<Msg> for Any {
     fn from(msg: Msg) -> Any {
         msg.0
-    }
-}
-
-impl From<AnyWrapper> for Any {
-    fn from(any_wrap: AnyWrapper) -> Any {
-        Any {
-            type_url: any_wrap.type_url,
-            value: any_wrap.value,
-        }
     }
 }
